move newsitem customization as all
